requirements.txt
xmltodict==0.13.0
Pillow==9.1.1
